Add .env.local for env vars
Adds a .env.local file into the root of the app, which can be picked by the dotenv gem if bundled into the app.

Now, any env vars that need to be defined for a rails app can be
defined as items in the `rails_app_user_envvars` variable.

Items are defined a dictionary with two required fields , envvar
and value, and when, which allows setting conditions for when the variable should be set.